BAU: Slight refactoring
Tidy up some files, apply formatting in some places, remove some unused commits
