+++ express/src/lib/lambda-facade/LambdaFacadeInterface.ts
@@ -6,10 +6,16 @@ import {AxiosResponse} from "axios";
 
 export default interface LambdaFacadeInterface {
     putUser(user: OnboardingTableItem, accessToken: string): Promise<AxiosResponse>;
+
     getUserByCognitoId(cognitoId: string, accessToken: string): Promise<AxiosResponse>;
+
     newService(service: Service, user: User, accessToken: string): Promise<AxiosResponse>;
+
     generateClient(serviceId: string, service: Service, contactEmail: string, accessToken: string): Promise<AxiosResponse>;
+
     updateClient(serviceId: string, selfServiceClientId: string, clientId: string, updates: object, accessToken: string): Promise<AxiosResponse>;
+
     listServices(userId: string, accessToken: string): Promise<AxiosResponse>;
+
     listClients(serviceId: string, accessToken: string): Promise<AxiosResponse>;
 }
