+++ express/src/lib/rfc822-validate/isRfc822Compliant.ts
@@ -13,6 +13,6 @@ let sAddrSpec = sLocalPart + '\\x40' + sDomain; // complete RFC822 email address
 let sValidEmail = '^' + sAddrSpec + '$'; // as whole string
 let reValidEmail = new RegExp(sValidEmail);
 
-export function isRfc822Compliant(email: string ): boolean {
+export function isRfc822Compliant(email: string): boolean {
     return reValidEmail.test(email);
 }
